API for getting historical file data created
