+++ data/urls.py
@@ -1,14 +1,16 @@
 from django.urls import path
 
-from . import views
+from .views import views, file_views, file_history_views
 
 app_name = "data"
 
 urlpatterns = [
-    path('upload_file/', views.upload_file),
-    path('get_file_data/', views.get_file_data),
-    path('get_file_names/', views.get_file_names),
-    path('update_file_data/', views.update_file_data),
+    path('upload_file/', file_views.upload_file),
+    path('get_file_data/', file_views.get_file_data),
+    path('get_file_history_names/', file_history_views.get_file_history_names),
+    path('get_history_file_data/', file_history_views.get_history_file_data),
+    path('get_file_names/', file_views.get_file_names),
+    path('update_file_data/', file_views.update_file_data),
 
     path('get_file_categories/', views.get_file_categories),
     path('add_new_category/', views.add_new_category),
